attempt at restoring SLO again

+++ src/SP.php
@@ -174,10 +174,80 @@ class SP
      */
     public function logout($relayState)
     {
+        if (false === $samlAssertion = $this->getAssertion()) {
+            // no session available
+            return $relayState;
+        }
+
+        $idpEntityId = $samlAssertion->getIssuer();
+        if (false === $idpInfo = $this->idpInfoSource->get($idpEntityId)) {
+            throw new SpException(\sprintf('IdP "%s" not registered', $idpEntityId));
+        }
         // delete the assertion, so we are no longer authenticated
         $this->session->delete('_fkooman_saml_sp_auth_assertion');
 
-        return $relayState;
+        $idpSloUrl = $idpInfo->getSloUrl();
+        if (null === $idpSloUrl) {
+            // IdP does not support SLO, nothing we can do about it
+            return $relayState;
+        }
+        if (null === $spSloUrl = $this->spInfo->getSloUrl()) {
+            // SP does not support SLO, do not redirect to IdP
+            return $relayState;
+        }
+
+        $requestId = \sprintf('_%s', Hex::encode($this->random->requestId()));
+        $logoutRequest = $this->tpl->render(
+            'LogoutRequest',
+            [
+                'ID' => $requestId,
+                'IssueInstant' => $this->dateTime->format('Y-m-d\TH:i:s\Z'),
+                'Destination' => $idpSloUrl,
+                'Issuer' => $this->spInfo->getEntityId(),
+                // we need the _exact_ (XML) NameID we got during
+                // authentication for the LogoutRequest
+                // XXX but it MUST be in the correct namespace, so we pretty much
+                // have to rewrite it to match LogoutRequest document namespace
+                'NameID' => $samlAssertion->getNameId(),
+            ]
+        );
+        $this->session->set('_fkooman_saml_sp_auth_logout_id', $requestId);
+        $this->session->set('_fkooman_saml_sp_auth_logout_idp', $idpEntityId);
+
+        return self::prepareRequestUrl($idpSloUrl, $logoutRequest, $relayState, $this->spInfo->getPrivateKey());
+    }
+
+    /**
+     * @param string $samlResponse
+     * @param string $relayState
+     * @param string $signature
+     *
+     * @return void
+     */
+    public function handleLogoutResponse($samlResponse, $relayState, $signature)
+    {
+        if (null === $spSloUrl = $this->spInfo->getSloUrl()) {
+            // SP does not support SLO, nothing we can do here...
+            return;
+        }
+
+        $idpEntityId = $this->session->get('_fkooman_saml_sp_auth_logout_idp');
+        if (false === $idpInfo = $this->idpInfoSource->get($idpEntityId)) {
+            throw new SpException(\sprintf('IdP "%s" not registered', $idpEntityId));
+        }
+
+        $logoutResponse = new LogoutResponse($this->dateTime);
+        $logoutResponse->verify(
+            $samlResponse,
+            $relayState,
+            $signature,
+            $this->session->get('_fkooman_saml_sp_auth_logout_id'),
+            $spSloUrl,
+            $idpInfo
+        );
+
+        $this->session->delete('_fkooman_saml_sp_auth_logout_id');
+        $this->session->delete('_fkooman_saml_sp_auth_logout_idp');
     }
 
     /**
@@ -206,6 +276,7 @@ class SP
                 'entityID' => $this->spInfo->getEntityId(),
                 'X509Certificate' => $this->spInfo->getPublicKey()->toEncodedString(),
                 'AssertionConsumerService' => $this->spInfo->getAcsUrl(),
+                'SingleLogoutService' => $this->spInfo->getSloUrl(),
             ]
         );
     }
